feat: move the request entity to it's own module
- the others remain in the same directory as they don't have an specific module
