+++ src/database/config/typeOrm.config.ts
@@ -1,20 +1,23 @@
-import { DataSource } from 'typeorm';
 import { ConfigService } from '@nestjs/config';
+import { TypeOrmModuleOptions } from '@nestjs/typeorm';
+import { DataSource, DataSourceOptions } from 'typeorm';
 import { config } from 'dotenv';
-import { resolve } from 'path';
+import { join } from 'path';
 
 config();
 
 const configService = new ConfigService();
 
-export default new DataSource({
+export const CONNECTION_OPTIONS: TypeOrmModuleOptions = {
   type: 'postgres',
   host: configService.get('DATABASE_HOST'),
   port: configService.get('DATABASE_PORT'),
   username: configService.get('DATABASE_USERNAME'),
   password: configService.get('DATABASE_PASSWORD'),
   database: configService.get('DATABASE_NAME'),
-  entities: [resolve('src', 'database/entities/*.entity{.ts,.js}')],
-  migrations: [resolve('src', 'database/migrations/*.ts')],
+  entities: [join(__dirname, '../entities', '*.entity.{js,ts}'), join(__dirname, '../..', 'api/**/entities/**', '*.entity.{js,ts}')],
+  migrations: [join(__dirname, '../migrations', '*.{js,ts}')],
   synchronize: false,
-});
+};
+
+export default new DataSource(CONNECTION_OPTIONS as DataSourceOptions);
