+++ src/database/entities/index.ts
@@ -4,3 +4,4 @@ export * from './driver.entity';
 export * from './location.entity';
 export * from './payment-method.entity';
 export * from './rider.entity';
+export * from './trip.entity';
