+++ src/database/index.ts
@@ -1,3 +1,3 @@
 export * from './config';
 export * from './database.module';
-export * from './services';
+export * from './entities';
